+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2023-07-27
+Logging Module Aristotle
+
+### Changed
+- Upgraded the Logging Module Aristotle:
+- (1) Functionality for Console or/and File Logging
+- (2) Restructuring of the Logging Utility to a Class (to manage multiple Logger Objects, Handlers, and Formatters)
+- (3) Fluid Options for setting Logging Thresholds
+- (4) Better Control on creation of Log Events
+- (5) Pre-Defined Log Events for restarts, cleanses, formatting, etc.
+
 ## [0.2.0] - 2023-07-24
 Creation of a simple CLI test utility to find an IP Address given an input hostname